hyperlight/requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
grpcio==1.68.1; python_version >= '3.8'
numpy==2.2.0; python_version >= '3.10'
pydantic==2.10.4; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
